03_my_implementations/chapter4_serving_patterns/05_prep_pred_pattern/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Machine Learning
onnxruntime==1.20.1
numpy==2.2.0
Pillow==11.0.0
joblib==1.4.2
scikit-learn==1.6.1

# gRPC and Protocol Buffers
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.3

# Data validation
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.0

# HTTP client
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
click==8.1.8
